Update root volume size on AMIs
This PR updates the root volume size on the AMIs that we create. This is necessary since the default size of 8GB is too small for our bare-metal instances to build QEMU images with.

This change should eventually be removed once we reconfigure the `arc-nv-runtime` to adjust the root device when launching the instance. This will allow us to save $$$ on the EBS snapshots associated with our AMIs.