ensure `compute-image-name.sh` fails if env vars aren't defined
`jq` will not throw an error if any of the `env.*` values are not defined. this can lead to issues since we expect these values to be defined.

to fix this, we can set the environment variables as `jq` arguments. this will cause the script to throw an error if the variables aren't defined due to the `set -u` option